requirements.txt
attrs==19.3.0
aws-sam-translator==1.15.1
beautifulsoup4==4.8.1
boto3==1.10.22
botocore==1.13.22
bs4==0.0.1
certifi==2019.9.11
cfn-lint==0.25.2
chardet==3.0.4
docutils==0.15.2
dukpy==0.2.2
idna==2.8
jmespath==0.9.4
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.2.0
numpy==1.18.4
pandas==1.0.3
pypac==0.13.0
pyrsistent==0.15.5
python-dateutil==2.8.0
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
tld==0.9.8
urllib3==1.25.7
